// src/api/generated/model/addressKindEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `physical` - Physical
 * * `shipping` - Shipping / Receiving
 * * `billing` - Billing
 * * `mailing` - Mailing
 */
export type AddressKindEnum = typeof AddressKindEnum[keyof typeof AddressKindEnum];


export const AddressKindEnum = {
  physical: 'physical',
  shipping: 'shipping',
  billing: 'billing',
  mailing: 'mailing',
} as const;